+++ src/server/graphql/mutations/addSlackAuth.ts
@@ -7,8 +7,68 @@ import standardError from '../../utils/standardError'
 import publish from 'server/utils/publish'
 import {TEAM} from 'universal/utils/constants'
 import {GQLContext} from 'server/graphql/graphql'
-import SlackNotification from 'server/database/types/SlackNotification'
+import SlackNotification, {SlackNotificationEvent} from 'server/database/types/SlackNotification'
 import SlackAuth from 'server/database/types/SlackAuth'
+
+const upsertNotifications = async (
+  viewerId: string,
+  teamId: string,
+  teamChannelId: string,
+  botChannelId: string
+) => {
+  const r = getRethink()
+  const existingNotifications = await r
+    .table('SlackNotification')
+    .getAll(viewerId, {index: 'userId'})
+    .filter({teamId})
+  const teamEvents = ['meetingStart', 'meetingEnd'] as SlackNotificationEvent[]
+  const userEvents = ['meetingStageTimeLimit'] as SlackNotificationEvent[]
+  const events = [...teamEvents, ...userEvents]
+  const upsertableNotifications = events.map((event) => {
+    const existingNotification = existingNotifications.find(
+      (notification) => notification.event === event
+    )
+    return new SlackNotification({
+      event,
+      // the existing notification channel could be a bad one (legacy reasons, bad means not public or not @Parabol)
+      channelId: teamEvents.includes(event) ? teamChannelId : botChannelId,
+      teamId,
+      userId: viewerId,
+      id: (existingNotification && existingNotification.id) || undefined
+    })
+  })
+  await r.table('SlackNotification').insert(upsertableNotifications, {conflict: 'replace'})
+}
+
+const upsertAuth = async (
+  viewerId: string,
+  teamId: string,
+  slackUserName: string,
+  slackRes: NonNullable<SlackManager['response']>
+) => {
+  const r = getRethink()
+  const existingAuth = (await r
+    .table('SlackAuth')
+    .getAll(viewerId, {index: 'userId'})
+    .filter({teamId})
+    .nth(0)
+    .default(null)) as SlackAuth | null
+  const slackAuth = new SlackAuth({
+    id: (existingAuth && existingAuth.id) || undefined,
+    createdAt: (existingAuth && existingAuth.createdAt) || undefined,
+    accessToken: slackRes.access_token,
+    teamId,
+    userId: viewerId,
+    slackTeamId: slackRes.team_id,
+    slackTeamName: slackRes.team_name,
+    slackUserId: slackRes.user_id,
+    slackUserName,
+    botUserId: slackRes.bot.bot_user_id,
+    botAccessToken: slackRes.bot.bot_access_token
+  })
+  await r.table('SlackAuth').insert(slackAuth, {conflict: 'replace'})
+  return slackAuth.id
+}
 
 export default {
   name: 'AddSlackAuth',
@@ -36,79 +96,31 @@ export default {
     }
 
     // RESOLUTION
-    const r = getRethink()
-
     const manager = await SlackManager.init(code)
     const {response} = manager
-    const {
-      access_token: accessToken,
-      bot,
-      team_id: slackTeamId,
-      team_name: slackTeamName,
-      user_id: slackUserId,
-      incoming_webhook: webhook
-    } = response
-
-    const userInfo = await manager.getUserInfo(slackUserId)
-    if (!userInfo.ok) {
-      return standardError(new Error(userInfo.error), {userId: viewerId})
+    const slackUserId = response.user_id
+    const defaultChannelId = response.incoming_webhook.channel_id
+    const [convoRes, userInfoRes, botChannelRes] = await Promise.all([
+      manager.getConversationInfo(defaultChannelId),
+      manager.getUserInfo(slackUserId),
+      manager.openIM(slackUserId)
+    ])
+    if (!userInfoRes.ok) {
+      return standardError(new Error(userInfoRes.error), {userId: viewerId})
     }
-    if (webhook) {
-      const existingTeamNotificationsCount = await r
-        .table('SlackNotification')
-        .getAll(teamId, {index: 'teamId'})
-        .filter({isActive: true})
-        .count()
-      if (existingTeamNotificationsCount === 0) {
-        // kick off some reasonable defaults if they're the first on the team to integration, else nothing
-        const {channel_id} = webhook
-        const notifications = [
-          new SlackNotification({
-            event: 'meetingStart',
-            channelId: channel_id,
-            teamId,
-            userId: viewerId
-          }),
-          new SlackNotification({
-            event: 'meetingEnd',
-            channelId: channel_id,
-            teamId,
-            userId: viewerId
-          })
-        ]
-        await r.table('SlackNotification').insert(notifications)
-      }
+    if (!botChannelRes.ok) {
+      return standardError(new Error(botChannelRes.error), {userId: viewerId})
     }
+    const {channel} = botChannelRes
+    const {id: botChannelId} = channel
+    // The default channel could be anything: public, private, im, mpim. Only allow public channels or the @Parabol channel
+    const teamChannelId = convoRes.ok ? defaultChannelId : botChannelId
 
-    const existingAuth = await r
-      .table('SlackAuth')
-      .getAll(viewerId, {index: 'userId'})
-      .filter({teamId})
-      .nth(0)
-      .default(null)
-    const slackAuth = new SlackAuth({
-      accessToken,
-      teamId,
-      userId: viewerId,
-      slackTeamId,
-      slackTeamName,
-      slackUserId,
-      slackUserName: userInfo.user.profile.display_name,
-      botUserId: bot.bot_user_id,
-      botAccessToken: bot.bot_access_token
-    })
-    if (existingAuth) {
-      slackAuth.id = existingAuth.id
-      slackAuth.createdAt = existingAuth.createdAt
-      await r
-        .table('SlackAuth')
-        .get(slackAuth.id)
-        .update(slackAuth)
-    } else {
-      await r.table('SlackAuth').insert(slackAuth)
-    }
-
-    const data = {slackAuthId: slackAuth.id, userId: viewerId}
+    const [, slackAuthId] = await Promise.all([
+      upsertNotifications(viewerId, teamId, teamChannelId, botChannelId),
+      upsertAuth(viewerId, teamId, userInfoRes.user.profile.display_name, response)
+    ])
+    const data = {slackAuthId, userId: viewerId}
     publish(TEAM, teamId, AddSlackAuthPayload, data, subOptions)
     return data
   }