+++ CHANGELOG.md
@@ -5,6 +5,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 This CHANGELOG follows conventions [outlined here](http://keepachangelog.com/).
 
+## vNEXT
+
+### Changed
+* Moved integration OAuth flow into their respective managers
+* Refactor all flow components to typescript
+
+### Removed
+* Legacy Provider mutations
+
 ## v3.13.0 2019-Jun-12
 
 ### Changed
@@ -21,7 +30,6 @@ This CHANGELOG follows conventions [outlined here](http://keepachangelog.com/).
 
 ### Changed
 * Moved integration OAuth flow into their respective managers
-* Refactor all flow components to typescript
 
 ## v3.12.0 2019-May-29